+++ product_code/user_service/CHANGELOG.md
@@ -1,8 +1,10 @@
-# [WIP - v0.14.2](https://github.com/upb-uc4/University-Credits-4.0/compare/user-v0.14.1...user-v0.14.2) (2020-XX-XX)
+# [WIP - v0.15.1](https://github.com/upb-uc4/University-Credits-4.0/compare/user-v0.14.1...user-v0.15.1) (2020-XX-XX)
 ## Feature
 ## Refactor
+- Added more precise logs for errors in topics
 - Enabled users to request the deletion of their accounts
 ## Bugfix
+- Added missing catch in user delete topic
 
 # [v0.14.1](https://github.com/upb-uc4/University-Credits-4.0/compare/user-v0.13.3...user-v0.14.1) (2020-12-14)
 ## Feature